demo/requirements.txt
Flask
gevent
greenlet
gunicorn
Pillow
tensorflow-gpu
